memento-agent: update for function-registry refactor
